+++ festival/src/main/java/com/example/festival/report/repository/ReportRepositoryInterface.java
@@ -3,6 +3,6 @@ package com.example.festival.report.repository;
 import com.example.festival.report.entity.Report;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReportRepositoryInterface extends CrudRepository<Report, Long> {
+public interface ReportRepositoryInterface extends CrudRepository<Report, Integer> {
 
 }
